third_party/blink/web_tests/external/wpt/css/css-ui/interactivity-inert-focus-expected.txt

This is a testharness.js-based test.
[FAIL] Focus non-inert input inside inert ancestry
  assert_equals: expected Element node <input id="inp_non_inert"></input> but got Element node <input id="inp_outer"></input>
Harness: the test ran to completion.
